Restructure Files
Create public/
Create public/javascript/
Move components/ to public/javascript/components
Move semantic/ to public/semantic/
Move favicons/ to public/favicons/
Move all .js files except for index.js and bundle.js to
public/javascript
Modify import statements to correctly account for new file structure.
