+++ index.html
@@ -3,8 +3,8 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="stylesheet" type="text/css" href="./semantic/dist/semantic.css">
-  <link rel="stylesheet" type="text/css" href="./semantic/dist/main.css">
+  <link rel="stylesheet" type="text/css" href="./public/semantic/dist/semantic.css">
+  <link rel="stylesheet" type="text/css" href="./public/semantic/dist/main.css">
   <title>OSSU</title>
