+++ index.js
@@ -4,8 +4,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import App from './components/App.js';
-import todoApp from './reducers';
+import App from './public/javascript/components/App.js';
+import todoApp from './public/javascript/reducers';
 
 let store = createStore(todoApp);
