+++ public/favicons/favicon.html
