+++ public/javascript/components/ContentBox.js
