+++ public/javascript/components/HorizontalAlignedRow.js
