+++ public/javascript/components/Masthead.js
