+++ public/javascript/components/NavBar.js
