+++ public/javascript/components/NavLogo.js
@@ -4,7 +4,7 @@ export class NavLogo extends React.Component {
   render () {
     return (
       <div href='#' className='navLogo header item'>
-        <img className='logo' src='./favicons/favicon-32x32.png' />
+        <img className='logo' src='./public/favicons/favicon-32x32.png' />
         OSSU
       </div>
     );
