+++ public/javascript/components/VerticalSegment.js
