+++ public/javascript/components/VerticalSegmentContent.js
